Split Trackable's print_* helpers and tap_* helpers specs
